[feat] 新增遠端MySQL的資料庫
1. 新增遠端MySQL的資料庫，未來無需再修改帳號密碼
