+++ src/main/java/com/lutu/ApiController.java
@@ -10,8 +10,8 @@ import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
-import com.lutu.camp_hibernate.model.CampService;
-import com.lutu.camp_hibernate.model.CampVO;
+import com.lutu.camp.model.CampService;
+import com.lutu.camp.model.CampVO;
 import com.lutu.campsite_order.model.CampSiteOrderService;
 import com.lutu.campsite_order.model.CampSiteOrderVO;
