+++ src/main/java/com/lutu/bundleitem/model/BundleItemService.java
@@ -6,8 +6,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.lutu.camp_hibernate.model.CampRepository;
-import com.lutu.camp_hibernate.model.CampVO;
 
 @Service("bundleItemService")
 public class BundleItemService {
