+++ src/main/java/com/lutu/camp/model/CampVO.java
@@ -1,4 +1,4 @@
-package com.lutu.camp_hibernate.model;
+package com.lutu.camp.model;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
